services/src/requirements.txt
fastapi
httpx
pydantic
python-dotenv
uvicorn
redis>=4.6.0
